Split Storager interface into Storager and ProjectStorager
